+++ models/goals.go
@@ -2,11 +2,11 @@ package models
 
 type Goal struct {
 	GormModel
-	Title       string `json:"title"`
-	Description string `json:"description"`
-	TargetValue int    `json:"target_value"`
-	Balance     int    `json:"balance"`
-	UserID      uint   `json:"user_id"`
-	Active      bool   `json:"active"`
-	Completed   bool   `json:"completed"`
+	Title       string  `json:"title"`
+	Description string  `json:"description"`
+	TargetValue float64 `json:"target_value"`
+	Balance     float64 `json:"balance"`
+	UserID      uint    `json:"user_id"`
+	Active      bool    `json:"active"`
+	Completed   bool    `json:"completed"`
 }
